requirements.txt
# Core ML / NLP
scikit-learn>=1.3,<2.0
pandas>=2.0,<3.0
nltk>=3.8,<4.0
numpy>=1.24,<2.0

# Web UI
streamlit>=1.30,<2.0

# Backend API
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.0,<3.0

# Testing
pytest>=8.0,<9.0
httpx>=0.27,<1.0

# Cloud Storage
azure-storage-blob>=12.0,<13.0
azure-identity>=1.15,<2.0
